Last classes remove, file deleted
