+++ Source/Frontend/UI/Components/Memory Tools/RTC_VmdAct_Form.cs
@@ -15,6 +15,21 @@
     using RTCV.Common.CustomExtensions;
     using static RTCV.UI.UI_Extensions;
 
+    [Serializable]
+    public class ActiveTableObject
+    {
+        public long[] Data { get; set; }
+
+        public ActiveTableObject()
+        {
+        }
+
+        public ActiveTableObject(long[] data)
+        {
+            Data = data;
+        }
+    }
+
     #pragma warning disable CA2213 //Component designer classes generate their own Dispose method
     public partial class RTC_VmdAct_Form : ComponentForm, IAutoColorize, IBlockable
     {
