+++ Source/Libraries/CorruptCore/INote.cs
@@ -1,0 +1,7 @@
+namespace RTCV.CorruptCore
+{
+    public interface INote
+    {
+        string Note { get; set; }
+    }
+}
