+++ Source/Libraries/CorruptCore/ProgressBar.cs
@@ -2,32 +2,6 @@ namespace RTCV.CorruptCore
 {
     using System;
 
-    [Serializable]
-    public class ActiveTableObject
-    {
-        public long[] Data { get; set; }
-
-        public ActiveTableObject()
-        {
-        }
-
-        public ActiveTableObject(long[] data)
-        {
-            Data = data;
-        }
-    }
-
-    public class ProblematicProcess
-    {
-        public string Name { get; set; }
-        public string Message { get; set; }
-    }
-
-    public interface INote
-    {
-        string Note { get; set; }
-    }
-
     public delegate void ProgressBarEventHandler(object source, ProgressBarEventArgs e);
 
     public class ProgressBarEventArgs : EventArgs
